Re-organize file upload controller
